Refactor code to apply packaging
