+++ src/main/java/exception/DukeDeadlineMissingDateException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeDeadlineMissingDateException extends DukeException {
