+++ src/main/java/exception/DukeDoneEmptyListException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeDoneEmptyListException extends DukeException {
