+++ src/main/java/exception/DukeEventMissingDateException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeEventMissingDateException extends DukeException {
