+++ src/main/java/exception/DukeMissingDescriptionException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeMissingDescriptionException extends DukeException {
