+++ src/main/java/exception/DukeNoSuchCommandException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeNoSuchCommandException extends DukeException {
