+++ src/main/java/exception/DukeSaveFailedException.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.exception;
 
 public class DukeSaveFailedException extends DukeException {
