+++ src/main/java/task/Event.java
@@ -1,27 +1,29 @@
-package seedu.duke;
+package seedu.duke.task;
 
+import seedu.duke.exception.DukeException;
+import seedu.duke.exception.DukeWrongDateFormatException;
 import java.util.Date;
 import java.text.SimpleDateFormat;
 import java.text.ParseException;
 
-public class Deadline extends Task {
+public class Event extends Task {
 
-    protected Date by;
+    private Date at;
     SimpleDateFormat dateFormatter = new SimpleDateFormat("dd/MM/yyyy HHmm");
 
-    public Deadline(String task, String by) throws DukeException {
+    public Event(String task, String at) throws DukeException {
         super(task);
         try {
-            this.by = dateFormatter.parse(by);
+            this.at = dateFormatter.parse(at);
         } catch (ParseException e) {
             throw new DukeWrongDateFormatException();
         }
     }
 
-    public Deadline(String task, String done, String by) throws DukeException {
+    public Event(String task, String done, String at) throws DukeException {
         super(task);
         try {
-            this.by = dateFormatter.parse(by);
+            this.at = dateFormatter.parse(at);
         } catch (ParseException e) {
             throw new DukeWrongDateFormatException();
         }
@@ -31,18 +33,18 @@ public class Deadline extends Task {
     }
 
     public String toStorageString() {
-        String output = "D|";
+        String output = "E|";
         if (super.isDone) {
             output = output + "1|";
         } else {
             output = output + "0|";
         }
-        output = output + super.taskName + "|" + dateFormatter.format(this.by);
+        output = output + super.taskName + "|" + dateFormatter.format(this.at);
         return output;
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (by: " + dateFormatter.format(by) + ")";
+        return "[E]" + super.toString() + " (at: " + dateFormatter.format(at) + ")";
     }
 }
