+++ src/main/java/task/Todo.java
@@ -1,4 +1,4 @@
-package seedu.duke;
+package seedu.duke.task;
 
 public class Todo extends Task {
